merge : S08P22D105-314-feat-12-be

See merge request s08-bigdata-recom-sub2/S08P22D105!71

Former-commit-id: b1d416b8fc819a5ed6667fe7f5499bf6c0e3a106

+++ backend/api-server/src/main/java/com/dawool/api/dto/detailInfo/EntertainmentDto.java
@@ -46,7 +46,7 @@ public class EntertainmentDto extends CommonInfoDto{
 
     @Builder
     public EntertainmentDto(String spotId, int contentId, int contentTypeId, String title, String category, String homepage, String firstImage,
-                            int areaCode, String addr1, float mapX, float mapY, float mLevel, int deaf, int visuallyImpaired, int mobilityWeak, int old, int infant, boolean isLiked, int hit, BarrierDto barrier,
+                            int areaCode, String addr1, double mapX, double mapY, float mLevel, int deaf, int visuallyImpaired, int mobilityWeak, int old, int infant, boolean isLiked, int hit, BarrierDto barrier,
                             int isBabyCarriage, int isPet, String expAgeRange, String expGuide, float heritage1, float heritage2, float heritage3,
                             String infoCenter, String commonParking, String restDate, String useSeason, String useTime) {
         super(spotId, contentId, contentTypeId, title, category, homepage, firstImage, areaCode, addr1, mapX, mapY, mLevel, deaf, visuallyImpaired, mobilityWeak, old, infant, isLiked, hit, barrier);
